pyproject.toml
[project]
name = "i-duruohacsc212"
version = "0.1.0"
description = ""
authors = [
    {name = "Duruoha Ikenna",email = "duruohaikenna@gmail.com"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "anyio (==4.9.0)",
    "appnope (==0.1.4)",
    "asttokens (==3.0.0)",
    "build (==1.2.2.post1)",
    "cachecontrol (==0.14.3)",
    "certifi (==2025.6.15)",
    "cffi (==1.17.1)",
    "charset-normalizer (==3.4.2)",
    "cleo (==2.1.0)",
    "comm (==0.2.2)",
    "contourpy (==1.3.2)",
    "crashtest (==0.4.1)",
    "cycler (==0.12.1)",
    "debugpy (==1.8.14)",
    "decorator (==5.2.1)",
    "distlib (==0.3.9)",
    "dulwich (==0.22.8)",
    "executing (==2.2.0)",
    "fastjsonschema (==2.21.1)",
    "filelock (==3.18.0)",
    "findpython (==0.6.3)",
    "fonttools (==4.58.4)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "installer (==0.7.0)",
    "ipykernel (==6.29.5)",
    "ipython (==9.3.0)",
    "ipython-pygments-lexers (==1.1.1)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.1.0)",
    "jedi (==0.19.2)",
    "jupyter-client (==8.6.3)",
    "jupyter-core (==5.8.1)",
    "keyring (==25.6.0)",
    "kiwisolver (==1.4.8)",
    "matplotlib (==3.10.3)",
    "matplotlib-inline (==0.1.7)",
    "more-itertools (==10.7.0)",
    "msgpack (==1.1.1)",
    "nest-asyncio (==1.6.0)",
    "numpy (==2.3.0)",
    "packaging (==25.0)",
    "parso (==0.8.4)",
    "pbs-installer (==2025.6.12)",
    "pexpect (==4.9.0)",
    "pillow (==11.2.1)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.3.8)",
    "poetry (==2.1.3)",
    "poetry-core (==2.1.3)",
    "prompt-toolkit (==3.0.51)",
    "psutil (==7.0.0)",
    "ptyprocess (==0.7.0)",
    "pure-eval (==0.2.3)",
    "pycparser (==2.22)",
    "pygame (==2.6.1)",
    "pygments (==2.19.1)",
    "pyopengl (==3.1.9)",
    "pyopengl-accelerate (==3.1.9)",
    "pyparsing (==3.2.3)",
    "pyproject-hooks (==1.2.0)",
    "python-dateutil (==2.9.0.post0)",
    "pyzmq (==27.0.0)",
    "rapidfuzz (==3.13.0)",
    "requests (==2.32.4)",
    "requests-toolbelt (==1.0.0)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "stack-data (==0.6.3)",
    "tomlkit (==0.13.3)",
    "tornado (==6.5.1)",
    "traitlets (==5.14.3)",
    "trove-classifiers (==2025.5.9.12)",
    "typing-extensions (==4.14.0)",
    "urllib3 (==2.4.0)",
    "virtualenv (==20.31.2)",
    "wcwidth (==0.2.13)",
    "xattr (==1.1.4)",
    "zstandard (==0.23.0)"
]

[tool.poetry]
packages = [{include = "i"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
